Add minimal compiler/bytecode tests
Bytecode testing is difficult for more complicated structures, like
loops and branches, because they depend on how the compiler chooses to
alter things - I believe that at that point, we should be testing
behavior and not bytecode output. The "push" and "pop" unit tests for
that section work well enough for now.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>