Add POPN instruction and syntax
There is a new VM instruction, POPN, which will pop the top N items off of the stack. This is so we don't have to do a bunch of separate lines of `.;` to discard multiple items. Also, it should be faster than multiple discards.

As a result of this change, the Action AST item is now an abstract class, with PopAction and PushAction taking its place. Although functionally the same, it's good to have two different types that represent two separate things.

Beyond that, the parser tests have been updated to reflect these changes.