Fix all ordered comparison builtin functions, alter `len` builtin behavior.
All ordered comparisions (<=, >=, <, >) were attempting to compare the
value wrappers, and not the values themselves in the functions that
would perform those operations.

The `len` fun no longer pops the item off of the stack. This is because
it's a simple readonly function that messes up the flow of messing with
things on the stack.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>